Skip registry config modification on locked SMB shares
Samba reads its running configuration from registry.tdb. Sharing.smb.update
will apply any share config updates to the registry, and so we need to
add a check for whether the share is locked. If the share is locked, then
we skip this step because it will not exist in the registry.