Fix race on failover with smb.configure / system dataset import
Due to nature of SMB service configuration dependencies, all service
initialization is done through the method smb.configure, which must
be performed after the configuration of the system dataset. This is
because required samba files are located on the system dataset, and
many steps must be performed in order. Base smb.conf must be genereated,
and then required samba directory tree created with specific files, then
our configuration synchronized with samba's registry.tdb.