Add legacy group-mapping capabilities
In legacy versions of TrueNAS, mappings for some builtin accounts
may have been allocated automatically by winbindd. Over time this
can result in unpredictable mappings, and if users rely on these
for granting access to paths in SMB shraes, it may result in
undefined behavior on upgrades. This patch allows support as a
workaround for when this has occured to set an auxiliary parameter
to trigger legacy group mappings for these accounts to be generated.