Remove Samba python binding usage from AD plugin
Attempting to initialize a samba3 loadparm context when ctdb
is disabled will cause an assertion / crash middleware. Work
around the issue by removing as many dependencies on python bindings
as possible. This had knock-on effects on other plugins as well.

Direct backend write methods were added to TDBWrap services
to allow unified way to bypass potentially complex logic of
normal methods for updating / changing the service configuration.
There are times that the extra work involved in a proper service
change is not desired.